refactor(bootstrap): Remove the `Sample` newtype
- The `Sample` newtype have been removed. Use `&[T]` instead.
- The one sample bootstrap is now available under the `Bootstrap::bootstrap`
  method
- The two sample bootstrap is now available under the `bootstrap` free function

Closes #16

[breaking-change]
